finished arith (?)

+++ arith/PixelStructs.h
@@ -34,7 +34,7 @@ typedef struct Dct_int {
 } *Dct_int;
 
 typedef struct Codeword {
-        uint64_t codeword;
+        uint32_t codeword;
 } *Codeword;
 
 float clampToRange(float input, float lower, float upper);
